backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
python-dotenv==1.1.1

openai==1.93.0
pinecone==7.3.0
langchain==0.3.26
langchain-core==0.3.67
langchain-openai==0.3.27
langchain-pinecone==0.2.8
langgraph==0.5.0
tiktoken==0.9.0

requests==2.32.4
